requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
sounddevice>=0.4.6
numpy>=1.21.0
wave
keyboard>=0.13.5
pyperclip>=1.8.2
PyQt6 